refactor(runway): split ratio geometry into runwayRatio.ts (R9 size gate)
runwayOfficial.ts hit 801>800 after rebuild on main. Extract the cohesive
ratio concern (per-model enums, orientation, video/image normalization duals)
— zero external users before split; runwayOfficial imports the three symbols
it consumes.
